refactor(step3) : product 패키지명 변경 -> products
