+++ src/test/java/kitchenpos/menus/application/MenuServiceTest.java
@@ -1,6 +1,5 @@
 package kitchenpos.menus.application;
 
-import kitchenpos.menus.application.MenuService;
 import kitchenpos.menus.domain.Menu;
 import kitchenpos.menus.domain.MenuGroupRepository;
 import kitchenpos.menus.domain.MenuProduct;
@@ -8,9 +7,9 @@ import kitchenpos.menus.domain.MenuRepository;
 import kitchenpos.menus.infrastructure.InMemoryMenuGroupRepository;
 import kitchenpos.menus.infrastructure.InMemoryMenuRepository;
 import kitchenpos.profanity.domain.PurgomalumClient;
-import kitchenpos.product.infrastructure.InMemoryProductRepository;
-import kitchenpos.product.domain.Product;
-import kitchenpos.product.domain.ProductRepository;
+import kitchenpos.products.infrastructure.InMemoryProductRepository;
+import kitchenpos.products.domain.Product;
+import kitchenpos.products.domain.ProductRepository;
 import kitchenpos.profanity.infrastructure.FakePurgomalumClient;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
